Svg load function to util file (move other similars?). Own svg used if climbing.
